move dir helpers from forc-util to sway-utils
