+++ forc-util/Cargo.toml
@@ -27,7 +27,6 @@ sway-utils = { version = "0.46.1", path = "../sway-utils" }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["ansi", "env-filter", "json"] }
 unicode-xid = "0.2.2"
-walkdir = "2.3.3"
 
 [features]
 default = ["fuel-tx"]
