+++ forc-util/src/lib.rs
@@ -9,14 +9,12 @@ use anyhow::{bail, Context, Result};
 use forc_tracing::{println_red_err, println_yellow_err};
 use std::{
     collections::{hash_map, HashSet},
-    str,
-};
-use std::{ffi::OsStr, process::Termination};
-use std::{
     fmt::Display,
     fs::File,
     hash::{Hash, Hasher},
     path::{Path, PathBuf},
+    process::Termination,
+    str,
 };
 use sway_core::language::parsed::TreeType;
 use sway_error::{
@@ -192,87 +190,6 @@ pub mod tx_utils {
             Ok(serde_json::to_string(&receipt_to_json_array)?)
         }
     }
-}
-
-/// Continually go down in the file tree until a Forc manifest file is found.
-pub fn find_nested_manifest_dir(starter_path: &Path) -> Option<PathBuf> {
-    find_nested_dir_with_file(starter_path, constants::MANIFEST_FILE_NAME)
-}
-
-/// Continually go down in the file tree until a specified file is found.
-///
-/// Starts the search from child dirs of `starter_path`.
-pub fn find_nested_dir_with_file(starter_path: &Path, file_name: &str) -> Option<PathBuf> {
-    use walkdir::WalkDir;
-    let starter_dir = if starter_path.is_dir() {
-        starter_path
-    } else {
-        starter_path.parent()?
-    };
-    WalkDir::new(starter_path)
-        .into_iter()
-        .filter_map(|e| e.ok())
-        .filter(|entry| entry.path() != starter_dir.join(file_name))
-        .filter(|entry| entry.file_name().to_string_lossy() == file_name)
-        .map(|entry| {
-            let mut entry = entry.path().to_path_buf();
-            entry.pop();
-            entry
-        })
-        .next()
-}
-
-/// Continually go up in the file tree until a specified file is found.
-///
-/// Starts the search from `starter_path`.
-#[allow(clippy::branches_sharing_code)]
-pub fn find_parent_dir_with_file<P: AsRef<Path>>(
-    starter_path: P,
-    file_name: &str,
-) -> Option<PathBuf> {
-    let mut path = std::fs::canonicalize(starter_path).ok()?;
-    let empty_path = PathBuf::from("/");
-    while path != empty_path {
-        path.push(file_name);
-        if path.exists() {
-            path.pop();
-            return Some(path);
-        } else {
-            path.pop();
-            path.pop();
-        }
-    }
-    None
-}
-/// Continually go up in the file tree until a Forc manifest file is found.
-pub fn find_parent_manifest_dir<P: AsRef<Path>>(starter_path: P) -> Option<PathBuf> {
-    find_parent_dir_with_file(starter_path, constants::MANIFEST_FILE_NAME)
-}
-
-/// Continually go up in the file tree until a Forc manifest file is found and given predicate
-/// returns true.
-pub fn find_parent_manifest_dir_with_check<T: AsRef<Path>, F>(
-    starter_path: T,
-    f: F,
-) -> Option<PathBuf>
-where
-    F: Fn(&Path) -> bool,
-{
-    find_parent_manifest_dir(starter_path).and_then(|manifest_dir| {
-        // If given check satisifies return current dir otherwise start searching from the parent.
-        if f(&manifest_dir) {
-            Some(manifest_dir)
-        } else if let Some(parent_dir) = manifest_dir.parent() {
-            find_parent_manifest_dir_with_check(parent_dir, f)
-        } else {
-            None
-        }
-    })
-}
-
-pub fn is_sway_file(file: &Path) -> bool {
-    let res = file.extension();
-    file.is_file() && Some(OsStr::new(constants::SWAY_EXTENSION)) == res
 }
 
 pub fn find_file_name<'sc>(manifest_dir: &Path, entry_path: &'sc Path) -> Result<&'sc Path> {
